avm2: Move supercalling logic to separate file
