+++ core/src/avm2/object/class_object.rs
@@ -2,18 +2,16 @@
 
 use crate::avm2::activation::Activation;
 use crate::avm2::class::{AllocatorFn, Class, CustomConstructorFn};
-use crate::avm2::error::{self, argument_error, make_error_1127, reference_error, type_error};
+use crate::avm2::error::{argument_error, make_error_1127, type_error};
 use crate::avm2::function::{exec, FunctionArgs};
 use crate::avm2::method::{Method, NativeMethodImpl};
 use crate::avm2::object::function_object::FunctionObject;
 use crate::avm2::object::script_object::ScriptObjectData;
 use crate::avm2::object::{Object, ScriptObject, TObject};
-use crate::avm2::property::Property;
 use crate::avm2::scope::{Scope, ScopeChain};
 use crate::avm2::value::Value;
 use crate::avm2::vtable::VTable;
 use crate::avm2::Error;
-use crate::avm2::Multiname;
 use crate::avm2::QName;
 use crate::avm2::TranslationUnit;
 use crate::string::AvmString;
@@ -351,262 +349,6 @@ impl<'gc> ClassObject<'gc> {
         } else {
             unreachable!("Cannot instantiate a class without an init method")
         }
-    }
-
-    /// Call a method defined in this class.
-    ///
-    /// This is intended to be called on the class object that is the
-    /// superclass of the one that defined the currently called property. If no
-    /// such superclass exists, you should use the class object for the
-    /// receiver's actual type (i.e. the lowest in the chain). This ensures
-    /// that repeated supercalls to the same method will call parent and
-    /// grandparent methods, and so on.
-    ///
-    /// If no method exists with the given name, this falls back to calling a
-    /// property of the `receiver`. This fallback only triggers if the property
-    /// is associated with a trait. Dynamic properties will still error out.
-    ///
-    /// This function will search through the class object tree starting from
-    /// this class up to `Object` for a method trait with the given name. If it
-    /// is found, it will be called with the receiver and arguments you
-    /// provided, as if it were defined on the target instance object.
-    ///
-    /// The class that defined the method being called will also be provided to
-    /// the `Activation` that the method runs on so that further supercalls
-    /// will work as expected.
-    ///
-    /// This method corresponds directly to the AVM2 operation `callsuper`,
-    /// with the caveat listed above about what object to call it on.
-    pub fn call_super(
-        self,
-        multiname: &Multiname<'gc>,
-        receiver: Object<'gc>,
-        arguments: FunctionArgs<'_, 'gc>,
-        activation: &mut Activation<'_, 'gc>,
-    ) -> Result<Value<'gc>, Error<'gc>> {
-        let property = self.instance_vtable().get_trait(multiname);
-        match property {
-            Some(Property::Slot { slot_id }) | Some(Property::ConstSlot { slot_id }) => {
-                let arguments = &arguments.to_slice();
-
-                let func = receiver.get_slot(slot_id);
-                func.call(activation, receiver.into(), arguments)
-            }
-            Some(Property::Method { disp_id }) => {
-                self.call_method_super(activation, receiver, disp_id, arguments)
-            }
-            Some(Property::Virtual { get: Some(get), .. }) => {
-                // Call the getter, then `Value::call` the result
-                let obj =
-                    self.call_method_super(activation, receiver, get, FunctionArgs::empty())?;
-
-                let arguments = &arguments.to_slice();
-                obj.call(activation, receiver.into(), arguments)
-            }
-            Some(Property::Virtual { get: None, .. }) => Err(error::make_reference_error(
-                activation,
-                error::ReferenceErrorCode::ReadFromWriteOnly,
-                multiname,
-                self.inner_class_definition(),
-            )),
-            None => {
-                let qualified_multiname_name = multiname.as_uri(activation.strings());
-                let qualified_class_name = self
-                    .inner_class_definition()
-                    .name()
-                    .to_qualified_name_err_message(activation.gc());
-
-                return Err(Error::avm_error(reference_error(
-                    activation,
-                    &format!(
-                        "Error #1070: Method {qualified_multiname_name} not found on {qualified_class_name}",
-                    ),
-                    1070,
-                )?));
-            }
-        }
-    }
-
-    /// Call a getter defined in this class.
-    ///
-    /// This is intended to be called on the class object that is the
-    /// superclass of the one that defined the currently called property. If no
-    /// such superclass exists, you should use the class object for the
-    /// receiver's actual type (i.e. the lowest in the chain). This ensures
-    /// that repeated supercalls to the same getter will call parent and
-    /// grandparent getters, and so on.
-    ///
-    /// If no getter exists with the given name, this falls back to getting a
-    /// property of the `receiver`. This fallback only triggers if the property
-    /// is associated with a trait. Dynamic properties will still error out.
-    ///
-    /// This function will search through the class object tree starting from
-    /// this class up to `Object` for a getter trait with the given name. If it
-    /// is found, it will be called with the receiver you provided, as if it
-    /// were defined on the target instance object.
-    ///
-    /// The class that defined the getter being called will also be provided to
-    /// the `Activation` that the getter runs on so that further supercalls
-    /// will work as expected.
-    ///
-    /// This method corresponds directly to the AVM2 operation `getsuper`,
-    /// with the caveat listed above about what object to call it on.
-    pub fn get_super(
-        self,
-        multiname: &Multiname<'gc>,
-        receiver: Object<'gc>,
-        activation: &mut Activation<'_, 'gc>,
-    ) -> Result<Value<'gc>, Error<'gc>> {
-        let property = self.instance_vtable().get_trait(multiname);
-
-        match property {
-            Some(Property::Slot { slot_id }) | Some(Property::ConstSlot { slot_id }) => {
-                Ok(receiver.get_slot(slot_id))
-            }
-            Some(Property::Method { disp_id }) => {
-                let full_method = self.instance_vtable().get_full_method(disp_id).unwrap();
-                let callee = FunctionObject::from_method(
-                    activation,
-                    full_method.method,
-                    full_method.scope(),
-                    Some(receiver.into()),
-                    full_method.super_class_obj,
-                    Some(full_method.class),
-                );
-
-                Ok(callee.into())
-            }
-            Some(Property::Virtual {
-                get: Some(disp_id), ..
-            }) => self.call_method_super(activation, receiver, disp_id, FunctionArgs::empty()),
-            Some(Property::Virtual { get: None, .. }) => Err(error::make_reference_error(
-                activation,
-                error::ReferenceErrorCode::ReadFromWriteOnly,
-                multiname,
-                self.inner_class_definition(),
-            )),
-            None => Err(error::make_reference_error(
-                activation,
-                error::ReferenceErrorCode::InvalidRead,
-                multiname,
-                self.inner_class_definition(),
-            )),
-        }
-    }
-
-    /// Call a setter defined in this class.
-    ///
-    /// This is intended to be called on the class object that is the
-    /// superclass of the one that defined the currently called property. If no
-    /// such superclass exists, you should use the class object for the
-    /// receiver's actual type (i.e. the lowest in the chain). This ensures
-    /// that repeated supercalls to the same setter will call parent and
-    /// grandparent setter, and so on.
-    ///
-    /// If no setter exists with the given name, this falls back to setting a
-    /// property of the `receiver`. This fallback only triggers if the property
-    /// is associated with a trait. Dynamic properties will still error out.
-    ///
-    /// This function will search through the class object tree starting from
-    /// this class up to `Object` for a setter trait with the given name. If it
-    /// is found, it will be called with the receiver and value you provided,
-    /// as if it were defined on the target instance object.
-    ///
-    /// The class that defined the setter being called will also be provided to
-    /// the `Activation` that the setter runs on so that further supercalls
-    /// will work as expected.
-    ///
-    /// This method corresponds directly to the AVM2 operation `setsuper`,
-    /// with the caveat listed above about what object to call it on.
-    pub fn set_super(
-        self,
-        multiname: &Multiname<'gc>,
-        value: Value<'gc>,
-        receiver: Object<'gc>,
-        activation: &mut Activation<'_, 'gc>,
-    ) -> Result<(), Error<'gc>> {
-        let property = self.instance_vtable().get_trait(multiname);
-        match property {
-            Some(Property::Slot { slot_id }) => {
-                // Comment placed for formatting reasons
-                receiver.set_slot(slot_id, value, activation)
-            }
-            Some(Property::Method { .. }) => Err(error::make_reference_error(
-                activation,
-                error::ReferenceErrorCode::AssignToMethod,
-                multiname,
-                self.inner_class_definition(),
-            )),
-            Some(Property::Virtual {
-                set: Some(disp_id), ..
-            }) => {
-                let args = FunctionArgs::AsArgSlice {
-                    arguments: &[value],
-                };
-
-                self.call_method_super(activation, receiver, disp_id, args)?;
-
-                Ok(())
-            }
-            Some(Property::ConstSlot { .. }) | Some(Property::Virtual { set: None, .. }) => {
-                if activation.is_interpreter() {
-                    Err(error::make_reference_error(
-                        activation,
-                        error::ReferenceErrorCode::WriteToReadOnly,
-                        multiname,
-                        self.inner_class_definition(),
-                    ))
-                } else {
-                    // In JIT mode in FP, setsuper on const slots and
-                    // getter-only accessors is silently ignored
-                    Ok(())
-                }
-            }
-            None => Err(error::make_reference_error(
-                activation,
-                error::ReferenceErrorCode::InvalidWrite,
-                multiname,
-                self.inner_class_definition(),
-            )),
-        }
-    }
-
-    /// Like `Value::call_method`, but uses the method specifically on this class's
-    /// instance vtable, not the instance vtable of the receiver. This is intended
-    /// to be used for supercalls.
-    fn call_method_super(
-        self,
-        activation: &mut Activation<'_, 'gc>,
-        receiver: Object<'gc>,
-        disp_id: u32,
-        arguments: FunctionArgs<'_, 'gc>,
-    ) -> Result<Value<'gc>, Error<'gc>> {
-        let full_method = self.instance_vtable().get_full_method(disp_id).unwrap();
-
-        // Only create callee if the method needs it
-        let callee = if full_method.method.needs_arguments_object() {
-            Some(FunctionObject::from_method(
-                activation,
-                full_method.method,
-                full_method.scope(),
-                Some(receiver.into()),
-                full_method.super_class_obj,
-                Some(full_method.class),
-            ))
-        } else {
-            None
-        };
-
-        exec(
-            full_method.method,
-            full_method.scope(),
-            receiver.into(),
-            full_method.super_class_obj,
-            Some(full_method.class),
-            arguments,
-            activation,
-            callee,
-        )
     }
 
     pub fn add_application(
